requirements.txt
requests
streamlit
openai
pinecone-client
